refactor: refactor database architecture to remove it from shared
