+++ src/preload/data.ts
@@ -1,0 +1,8 @@
+export const data = {
+  targetLanguage: "eng",
+  // temp for development
+  station: {
+    name: "5-temple-fork",
+    id: 5
+  }
+};
