+++ src/preload/index.ts
@@ -1,22 +1,17 @@
 import { electronAPI } from "@electron-toolkit/preload";
 import { contextBridge } from "electron";
+import { data } from "./data";
 
-// Custom APIs for renderer
-const api = {};
-
-// Use `contextBridge` APIs to expose Electron APIs to
-// renderer only if context isolation is enabled, otherwise
-// just add to the DOM global.
 if (process.contextIsolated) {
   try {
     contextBridge.exposeInMainWorld("electron", electronAPI);
-    contextBridge.exposeInMainWorld("api", api);
+    contextBridge.exposeInMainWorld("data", data);
   } catch (error) {
     console.error(error);
   }
 } else {
   // @ts-ignore (define in dts)
-  window.electron = electronAPI;
+  window.data = data;
   // @ts-ignore (define in dts)
-  window.api = api;
+  window.electron = electronAPI;
 }
